move cmd_test from cmd.c to log.c, not test yet
